requirements.txt
shiny
pandas
plotly-express
shinywidgets
shinyswatch
google-cloud-storage
fsspec
gcsfs


